requirement.txt

black==22.3.0
pytest==7.4.0
pytest-cov==4.0.0
pylint==2.15.3

fastapi==0.85.0         
uvicorn==0.18.3         

pandas==1.5.3
databricks-sql-connector==3.4.0

python-dotenv==1.0.1
